CRAS: linear_resampler - increase resample precision
When doing SRC, the actual rate is trimmed to the floor
integer before use and causing a precision lost. Multiply
the numerator and denominator of the rate factor by 100,
so it takes more than 1 day for the lost frames count to be
accumulated to 1024.

BUG=chromium:420589
TEST=Play youtube to two active devices

Reviewed-on: https://chromium-review.googlesource.com/221706
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>

+++ cras/src/server/cras_fmt_conv.c
@@ -608,8 +608,8 @@ size_t cras_fmt_conv_out_frames_to_in(struct cras_fmt_conv *conv,
 }
 
 void cras_fmt_conv_set_linear_resample_rates(struct cras_fmt_conv *conv,
-					     unsigned int from,
-					     unsigned int to)
+					     float from,
+					     float to)
 {
 	linear_resampler_set_rates(conv->resampler, from, to);
 }
