+++ cras/src/server/cras_fmt_conv.h
@@ -41,8 +41,8 @@ size_t cras_fmt_conv_out_frames_to_in(struct cras_fmt_conv *conv,
 				      size_t out_frames);
 /* Sets the input and output rate to the linear resampler. */
 void cras_fmt_conv_set_linear_resample_rates(struct cras_fmt_conv *conv,
-					     unsigned int from,
-					     unsigned int to);
+					     float from,
+					     float to);
 /* Converts in_frames samples from in_buf, storing the results in out_buf.
  * Args:
  *    conv - The format converter returned from cras_fmt_conv_create().
